requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
neo4j==5.28.1
nltk==3.9.1
openai==1.70.0
pydantic==2.11.2
pydantic_core==2.33.1
PyPDF2==3.0.1
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
